--- src/components/shared/custom-modal/index.js
@@ -1,3 +1,0 @@
-import CustomModal from './custom-modal';
-
-export default CustomModal;
