+++ src/components/shared/form-modal/index.js
@@ -1,0 +1,3 @@
+import FormModal from './form-modal';
+
+export default FormModal;
